Incorporate safer/more flexible URL handling in TeamCity plugin
For cases in which a Phabricator installation runs on a custom port or doesn't support plain HTTP, this plugin is currently limited. This modifies the phabricatorUrl in the AppConfig to be a URL rather than a String, which makes the TeamCity build feature configuration more flexible. I've built and tested this change within my local CI environment.